pyproject.toml
[tool.poetry]
name = "capstone-project"
version = "0.1.0"
description = "Data-Driven Software Security Assessment"
authors = ["Samuel Gamelin <samuel.gamelin@carleton.ca>", "John Breton <john.breton@carleton.ca>", "Mohamed Radwan <mohamed.radwan@carleton.ca>", "Khalil Aalab <khalil.aalab@carleton.ca>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
PyQt5 = "^5.15.6"
pandas = "^1.3.5"
requirements-parser = "^0.3.1"
setup-py-upgrade = "^1.2.1"
setup-cfg-fmt = "^1.20.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
